refactor the tests into two suites - existing default and one with Sessions enabled
